Before implementing Inverse and determinant

+++ A3_2020274/src/ColumnMatrix.java
@@ -17,7 +17,7 @@ public class ColumnMatrix extends RectangularMatrix{
     }
 
     @Override
-    public int determinant() {
+    public double determinant() {
         return 0;
     }
