+++ A3_2020274/src/TriangularMatrix.java
@@ -8,7 +8,7 @@ public class TriangularMatrix extends SquareMatrix {
 
     //TODO: check if this works for all the lower classes
     @Override
-    public int determinant() {
+    public double determinant() {
         int temp = 1;
         for (int i = 0; i < this.m.length; i++) {
             temp *= this.m[i][i];
